+++ extensions/launchdarkly/src/switch-project.ts
@@ -1,3 +1,0 @@
-import SwitchProject from "./components/SwitchProject";
-
-export default SwitchProject;